+++ src/main/java/org/folio/marc4ld/service/marc2ld/bib/MarcBib2LdMapperImpl.java
@@ -84,8 +84,8 @@ public class MarcBib2LdMapperImpl implements MarcBib2ldMapper {
   private Resource createInstanceAndWorkResource(List<Record> records) {
     var instanceAndWork = createInstanceAndWork();
     records.forEach(marcRecord -> fillInstanceFields(marcRecord, instanceAndWork.instance));
-    setAdditionInfo(instanceAndWork);
     emptyEdgesCleaner.apply(instanceAndWork.instance);
+    setAdditionInfo(instanceAndWork);
     return instanceAndWork.instance;
   }
 
@@ -120,7 +120,7 @@ public class MarcBib2LdMapperImpl implements MarcBib2ldMapper {
         rules.findBibFieldRules(tag)
           .stream()
           .filter(rule -> conditionChecker
-            .isMarc2LdConditionSatisfied(rule.getOriginal(), dataField, marcRecord.getControlFields()))
+            .isMarc2LdConditionSatisfied(rule.getOriginal(), field, marcRecord.getControlFields()))
           .forEach(fr -> fieldController.handleField(instance, field, marcRecord.getControlFields(), fr))
       );
   }