+++ NEWS.md
@@ -1,3 +1,11 @@
+## 2.0.0 (IN PROGRESS)
+- Update Mapping for 300$b & $e (Physical description & Accompanying material) [MODLD-508](https://folio-org.atlassian.net/browse/MODLD-508)
+- Convert marc field 545 to Resource description [MODLD-337](https://folio-org.atlassian.net/browse/MODLD-337)
+- Derive MARC 300 from Graph [MODLD-401](https://folio-org.atlassian.net/browse/MODLD-401)
+- Convert marc field 586 (Awards Note) to Resource description [MODLD-366](https://folio-org.atlassian.net/browse/MODLD-366)
+- Create Place resource from marc auth field 151 when $v, $x, $y, $z not present [MODLD-529](https://folio-org.atlassian.net/browse/MODLD-529)
+- Convert marc field 041 (Language categories) to Resource description [MODLD-361](https://folio-org.atlassian.net/browse/MODLD-361)
+
 ## 1.0.3 (04-09-2025)
 - Geographic coverage: single "geographicCoverage" edge is created for Work which contains 043 with multiple $a fields [MODLD-694](https://folio-org.atlassian.net/browse/MODLD-694)
 - "targetAudience", "geographicCoverage" links contain https instead of http [MODLD-695](https://folio-org.atlassian.net/browse/MODLD-695)